requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.121.0
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uuid==1.30
uvicorn==0.38.0
bcrypt
python-multipart
langchain_openai
langchain-community
langchain
langgraph
langchain_text_splitters
python-docx
pypdf